Add context manager
The context manager serves as the center of the UI -- whenever a context
gets activated, it will serve as the basis for sending out signals
informing the rest of the UI that a context has been activated and
giving it the chance to respond accordingly.

QtContexts, when initialized, register themselves with the manager.
When destroyed, they unregister themselves.